refactor: extract TerrainType to domain values layer
- Move TerrainType from map.rs to domain/values/terrain.rs
- Update all imports across the codebase
- TerrainType now properly separated as value object
- All terrain logic centralized in values layer
