+++ src/domain/values/mod.rs
@@ -1,1 +1,2 @@
-pub mod resource; 
+pub mod resource;
+pub mod terrain;
