+++ src/simulation/map.rs
@@ -1,30 +1,11 @@
 use crate::simulation::entities::{Map, ResourceType};
+use crate::domain::values::terrain::TerrainType;
 use noise::{NoiseFn, Perlin};
 use rand::prelude::*;
 use rand_chacha::ChaCha8Rng;
 use std::collections::HashMap;
 use std::fs;
 use std::path::Path;
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-pub enum TerrainType {
-    Plain = 0,
-    Hill = 1,
-    Mountain = 2,
-    Canyon = 3,
-}
-
-impl From<u8> for TerrainType {
-    fn from(value: u8) -> Self {
-        match value {
-            0 => TerrainType::Plain,
-            1 => TerrainType::Hill,
-            2 => TerrainType::Mountain,
-            3 => TerrainType::Canyon,
-            _ => TerrainType::Plain,
-        }
-    }
-}
 
 pub struct MapConstants {
     pub noise_frequency: f64,
@@ -151,16 +132,4 @@ impl Map {
     }
 }
 
-impl TerrainType {
-    pub fn is_traversable(&self) -> bool {
-        matches!(self, TerrainType::Plain | TerrainType::Hill)
-    }
 
-    pub fn movement_cost(&self) -> u32 {
-        match self {
-            TerrainType::Plain => 1,
-            TerrainType::Hill => 2,
-            TerrainType::Mountain | TerrainType::Canyon => 0,
-        }
-    }
-}
