+++ src/simulation/robot_ai/pathfinding.rs
@@ -1,5 +1,5 @@
 use crate::simulation::entities::Map;
-use crate::simulation::map::TerrainType;
+use crate::domain::values::terrain::TerrainType;
 use crate::simulation::robot_ai::types::Direction;
 use std::cmp::Ordering;
 use std::collections::{BinaryHeap, HashMap};
